refactor: services 폴더 구조 수정(client, server분리)
